--- backend/src/main/java/com/festago/auth/domain/OpenIdUserInfoProvider.java
@@ -1,6 +1,0 @@
-package com.festago.auth.domain;
-
-public interface OpenIdUserInfoProvider {
-
-    UserInfo provide(String idToken);
-}
